Add files via upload
This is the extra-credit skipped list program. I was late because of the paramount amount of pressure of the finals and hence took a long time before could be pushed here. I hope that minimal marks are deducted for that as at Ashoka academic pressure is quite high as you are aware.